+++ changelog.md
@@ -1,3 +1,8 @@
+# 0.0.3
+
+* Added drill control to the Tunnel Digger's GUI
+* Added sonar to the Copperfins GUI
+
 # 0.0.2
 
 * Fixed crashes with dedicated servers and forge